change models & templates & forms & views; add categories fixtures

+++ recipebook/recipebook/apps/recipe/views.py
@@ -2,7 +2,8 @@ import datetime
 from django.http import Http404, HttpResponseForbidden, HttpResponseRedirect
 from django.shortcuts import redirect, render, get_object_or_404
 from django.urls import reverse
-from .models import Recipe, UserProfile
+from .models import Recipe #, UserProfile
+from django.contrib.auth.models import User
 from .forms import RecipeForm, EditRecipeForm, SignUpForm
 from django.contrib.auth.forms import AuthenticationForm
 from django.contrib.auth import authenticate, login, logout
@@ -18,14 +19,20 @@ def index(request):
 
 
 def detail(request, recipe_id):
-    try:
-        r = Recipe.objects.get(id = recipe_id)
-    except:
-        raise Http404("!Рецепт не найден!")
-    latest_comments_list = r.comment_set.order_by("-id")[:5]
+    # try:
+    #     r = Recipe.objects.get(id = recipe_id)
+    # except:
+    #     raise Http404("!Рецепт не найден!")
+    recipe = get_object_or_404(Recipe, id = recipe_id)
+    latest_comments_list = recipe.comment_set.order_by("-id")[:5]
+    if request.user == recipe.author:
+        auth_state = True
+    else:
+        auth_state = False
     context = {
-        'recipes': r,
-        'latest_comments_list':latest_comments_list
+        'recipes': recipe,
+        'latest_comments_list':latest_comments_list,
+        'auth_state': auth_state
     }
     return render(request,'recipe/detail.html', context)
 
@@ -45,26 +52,14 @@ def add_recipe(request):
         form = RecipeForm(request.POST, request.FILES)
         if form.is_valid():
             recipe = form.save(commit=False)
-            profile, _ = UserProfile.objects.get_or_create(user=request.user)
-            recipe.author = profile
-            recipe.pub_date = datetime.datetime.now()
+            recipe.author = request.user            # Связываем автора рецепта с текущим пользователем
+            # recipe.pub_date = datetime.datetime.now()
             recipe.save()
             messages.success(request, 'Рецепт успешно добавлен!')
             return HttpResponseRedirect(reverse_lazy('recipes:index'))
     else:
         form = RecipeForm()
     return render(request, 'recipe/add_recipe.html', {'form': form})
-# def add_recipe(request):
-#     if request.method == 'POST':
-#         form = RecipeForm(request.POST, request.FILES)
-#         if form.is_valid():
-#             recipe = form.save(commit=False)
-#             recipe.author = request.user.profile
-#             recipe.save()
-#             return redirect('recipe/book.html')
-#     else:
-#         form = RecipeForm()
-#     return render(request, 'recipe/add_recipe.html', {'form': form})
 
 
 def edit_recipe(request, recipe_id):
@@ -79,8 +74,12 @@ def edit_recipe(request, recipe_id):
 def delete_recipe(request, recipe_id):
     recipe = get_object_or_404(Recipe, id = recipe_id)
     if recipe.author == request.user:
-        recipe.delete()
-        return HttpResponseRedirect(reverse_lazy('recipes:index'))
+        if request.method == 'POST':
+            # Удаляем рецепт
+            recipe.delete()
+            # Django автоматически удалит фотографию, связанную с рецептом
+            return HttpResponseRedirect(reverse_lazy('recipes:index'))  # Переадресация на список рецептов
+        return render(request, 'recipe/delete_recipe.html', {'recipe': recipe})
     else:
         return HttpResponseForbidden()
     
@@ -99,16 +98,26 @@ class RegisterUser(CreateView):
         return valid
 
 # Авторизация пользователя
+def check_if_user_exists(username):
+    """
+    Функция проверяет, существует ли пользователь с указанным именем.
+    :param username: имя пользователя для проверки
+    :return: True, если пользователь существует, False в противном случае
+    """
+    return User.objects.filter(username=username).exists()
+
 def login_user(request):
     if request.method == 'POST':
         form = AuthenticationForm(data=request.POST)
         if form.is_valid():
-            username = form.cleaned_data.get('username')
             password = form.cleaned_data.get('password')
-            user = authenticate(username=username, password=password)
-            if user is not None:
-                login(request, user)
-                return HttpResponseRedirect(reverse_lazy('recipes:index'))
+            # Получаем имя пользователя из запроса
+            username = request.POST.get('username', '')
+            if check_if_user_exists(username):
+                user = authenticate(username=username, password=password)
+                if user is not None:
+                    login(request, user)
+                    return HttpResponseRedirect(reverse_lazy('recipes:index'))
     else:
         form = AuthenticationForm()
     return render(request, 'recipe/login.html', {'form': form})
